+++ zokrates_core/src/lib.rs
@@ -7,4 +7,3 @@ mod macros;
 mod optimizer;
 mod semantics;
 mod static_analysis;
-pub mod utils;
