--- zokrates_core/src/utils/mod.rs
@@ -1,1 +1,0 @@
-pub mod macros;
